\name{print.Bain}
\alias{print.Bain}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Bain result print.
}
\description{
  This function prints result of Bain functions.
}
\usage{
\method{print}{Bain}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    Result from Bain, Bain_ttest, Bain_ttestData, Bain_anova, Bain_anova_core,
    Bain_ancova, Bain_ancova_core or Bain_regression function.
  }
  \item{...}{
    Optional arguments passed to the print function
  }
}

\seealso{
  \code{\link{Bain},\link{print}}
}

\alias{print.Bain}



